# packages/core/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TeachingCommand": {
        "additionalProperties": false,
        "properties": {
          "arguments": {
            "additionalProperties": true,
            "title": "Arguments",
            "type": "object"
          },
          "command_id": {
            "pattern": "^[A-Za-z0-9_-]{1,96}$",
            "title": "Command Id",
            "type": "string"
          },
          "episode_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9_-]{1,96}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Id"
          },
          "expected_revision": {
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "operation": {
            "enum": [
              "task",
              "start",
              "pause",
              "reset",
              "correct",
              "mark_failure",
              "finish"
            ],
            "title": "Operation",
            "type": "string"
          },
          "session_id": {
            "pattern": "^[A-Za-z0-9_-]{1,96}$",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "command_id",
          "expected_revision",
          "operation"
        ],
        "title": "TeachingCommand",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VoiceJoin": {
        "additionalProperties": false,
        "properties": {
          "session_id": {
            "pattern": "^[A-Za-z0-9_-]{1,96}$",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "VoiceJoin",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Firebird local VLA application",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/teaching/commands": {
      "post": {
        "operationId": "command_api_v1_teaching_commands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeachingCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Command",
        "tags": [
          "Teaching"
        ]
      }
    },
    "/api/v1/teaching/commands/{command_id}": {
      "get": {
        "operationId": "receipt_api_v1_teaching_commands__command_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "command_id",
            "required": true,
            "schema": {
              "title": "Command Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Receipt",
        "tags": [
          "Teaching"
        ]
      }
    },
...
    "/api/v1/teaching/voice/join": {
      "post": {
        "operationId": "voice_join_api_v1_teaching_voice_join_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceJoin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Voice Join",
        "tags": [
          "Teaching"
        ]
      }
    }
  }
}